Fixed the behavior of backspace key when using search_key=None
When search_key=None, if we press backspace as the first character, the
word is entered literally ("backspace") as the search term.

If the search_key is None and backspace is pressed, do nothing. (Do not
set the search_key to "backspace".)

Partially resolves https://github.com/IngoMeyer441/simple-term-menu/issues/41
(The issue with accented characters mentioned in the issue description is not
handled.)